SLING-8104 Avoid magic when merging features
Handle artifact includes without the need for declaring external overrides.
When an artifact with the same groupID and artifactID but a different
version is specified in the feature that includes another feature, the one
from the including feature is always selected, i.e. the included one is
overridden.
This commit adds one new override specification mechanism. To apply the same
override rule for all clashes, a wildcard using '*' for groupID and
artifactID can be used, e.g:
 *:*:LATEST
means always select the latest in case of a clash.