SLING-8104 Avoid magic when merging features
Support for side-by-side versions of artifacts when inheriting.
When a feature inherits from another feature that lists the same
artifact twice in different versions, both are inherited, unless the
target feature also specifies one or more artifacts with the same group
ID and artifact ID, in that case the artifacts from the included
features are replaced with the one(s) specified in the target feature.