SLING-8104 Avoid magic when merging features
When merging artifacts/bundles, they need to be selected from a provided
override list if the artifact versions are not the same. The list has
the following syntax:
  groupid1:artifactid1:<resolution>
  groupid2:artifactid2:<resolution>
To apply the same override rule for all clashes, a wildcard using '*' for
groupID and artifactID can be used:
  *:*:<resolution>
means always select the same resolution in case of a clash.

Where <resolution> is one of the following:
  ALL - select all the artifacts
  HIGHEST - select only the artifact with the highest version number
  LATEST - select only the artifact provided latest
  <version> - select this specific version
When comparing version numbers these are converted to OSGi version
numbers and the OSGi version number ordering is applied.

When merging includes artifacts specified in the target feature
override all artifacts with the same group ID and artifact ID in the
included feature.
Both the included as well the target feature can have multiple artifacts
with the same group ID and artifact ID but different versions.